first batch of refactoring before clean up
