+++ great_expectations/data_context/data_context/file_data_context.py
@@ -1,5 +1,9 @@
+import errno
 import logging
-from typing import Mapping, Optional, Union
+import os
+from typing import Mapping, Optional, Union, cast
+
+from ruamel.yaml import YAML
 
 from great_expectations.data_context.data_context.abstract_data_context import (
     AbstractDataContext,
@@ -8,8 +12,14 @@ from great_expectations.data_context.types.base import DataContextConfig
 from great_expectations.data_context.types.data_context_variables import (
     FileDataContextVariables,
 )
+from great_expectations.data_context.util import substitute_config_variable
 
 logger = logging.getLogger(__name__)
+
+# TODO: check if this can be refactored to use YAMLHandler class
+yaml = YAML()
+yaml.indent(mapping=2, sequence=4, offset=2)
+yaml.default_flow_style = False
 
 
 class FileDataContext(AbstractDataContext):
@@ -36,6 +46,11 @@ class FileDataContext(AbstractDataContext):
                 config_variables.yml and the environment
         """
         super().__init__(runtime_environment=runtime_environment)
+        # move this down here too
+        if context_root_dir is not None:
+            context_root_dir = os.path.abspath(context_root_dir)
+        self._context_root_directory = context_root_dir
+
         self._context_root_dir = context_root_dir
         self._project_config = self._apply_global_config_overrides(
             config=project_config
@@ -43,3 +58,42 @@ class FileDataContext(AbstractDataContext):
 
     def _init_variables(self) -> FileDataContextVariables:
         raise NotImplementedError
+
+    def _load_config_variables(self) -> dict:
+        """
+        Get all config variables from the default location. For Data Contexts in GE Cloud mode, config variables
+        have already been interpolated before being sent from the Cloud API.
+        """
+        print("runningg me!!!")
+        config_variables_file_path = cast(
+            DataContextConfig, self.get_config()
+        ).config_variables_file_path
+        if config_variables_file_path:
+            try:
+                # If the user specifies the config variable path with an environment variable, we want to substitute it
+                defined_path = substitute_config_variable(
+                    config_variables_file_path, dict(os.environ)
+                )
+                if not os.path.isabs(defined_path):
+                    # TODO: is this something that needs to be revisited?
+                    # A BaseDataContext will not have a root directory; in that case use the current directory
+                    # for any non-absolute path
+                    root_directory = self.root_directory or os.curdir
+                else:
+                    root_directory = ""
+                var_path = os.path.join(root_directory, defined_path)
+                with open(var_path) as config_variables_file:
+                    return yaml.load(config_variables_file) or {}
+            except OSError as e:
+                if e.errno != errno.ENOENT:
+                    raise
+                logger.debug("Generating empty config variables file.")
+                return {}
+        else:
+            return {}
+
+    @property
+    def root_directory(self):
+        """The root directory for configuration objects in the data context; the location in which
+        ``great_expectations.yml`` is located."""
+        return self._context_root_directory
